lima-init: work around invalid sshfs mounts
Lima produces invalid `user-data` configuration when using sshfs with
rosetta; this causes the init script to semi-silently fail halfway and not
set up networking correctly.  Work around the issue by manually checking
for `sshfs` and skipping them.

A proper fix will be made in upstream lima to not generate erroneous
configuration in the first place.

Signed-off-by: Mark Yen <mark.yen@suse.com>